requirements.txt
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
gunicorn==20.1.0
psycopg2==2.9.3
whitenoise==6.2.0